test(trino): move pure parser tests into tests/unit/ so CI runs them
The container-backed test_trino.py is not exercised by any CI job
(test-unit runs tests/unit/, test-connector only covers postgres+mysql),
so the URL-decode regression tests never ran. Split the pure
_parse_trino_url / _parse_trino_data_type / _build_trino_column /
_strip_trailing_semicolon / lazy-import tests into
tests/unit/test_trino_parser.py (no testcontainers import) so the
test-unit job collects and runs them. The container suite stays in
tests/connectors/test_trino.py.
